+++ src/Giraffe.Razor/HttpHandlers.fs
@@ -7,6 +7,7 @@ module HttpHandlers =
     open System.Threading.Tasks
     open System.Collections.Generic
     open Microsoft.AspNetCore.Http
+    open Microsoft.AspNetCore.Mvc.ModelBinding
     open Microsoft.AspNetCore.Mvc.Razor
     open Microsoft.AspNetCore.Mvc.ViewFeatures
     open Microsoft.Extensions.DependencyInjection
@@ -20,12 +21,13 @@ module HttpHandlers =
     let razorView (contentType : string)
                   (viewName    : string)
                   (model       : 'T option)
-                  (viewData    : IDictionary<string, obj> option) : HttpHandler =
+                  (viewData    : IDictionary<string, obj> option)
+                  (modelState  : ModelStateDictionary option) : HttpHandler =
         fun (next : HttpFunc) (ctx : HttpContext) ->
             task {
                 let engine = ctx.RequestServices.GetService<IRazorViewEngine>()
                 let tempDataProvider = ctx.RequestServices.GetService<ITempDataProvider>()
-                let! result = renderView engine tempDataProvider ctx viewName model viewData
+                let! result = renderView engine tempDataProvider ctx viewName model viewData modelState
                 match result with
                 | Error msg -> return (failwith msg)
                 | Ok output ->
@@ -35,10 +37,11 @@ module HttpHandlers =
 
     /// Reads a razor view from disk and compiles it with the given model and sets
     /// the compiled output as the HTTP reponse with a Content-Type of text/html.
-    let razorHtmlView (viewName : string)
-                      (model    : 'T option)
-                      (viewData : IDictionary<string, obj> option) : HttpHandler =
-        razorView "text/html; charset=utf-8" viewName model viewData
+    let razorHtmlView (viewName   : string)
+                      (model      : 'T option)
+                      (viewData   : IDictionary<string, obj> option)
+                      (modelState : ModelStateDictionary option) : HttpHandler =
+        razorView "text/html; charset=utf-8" viewName model viewData modelState
 
     /// Validates an anti forgery token.
     /// If the token is valid the handler will procceed as normal,